--- src/statd/python/yanger/__init__.py
@@ -1,4 +1,0 @@
-from .yanger import main
-
-if __name__ == "__main__":
-    main()
