+++ src/statd/python/yanger/__main__.py
@@ -1,0 +1,3 @@
+from .yanger import main
+
+main()
