+++ test/case/cli/system-output/run/chronyc_-c_sources
