+++ test/case/cli/system-output/run/fw_printenv_BOOT_ORDER
