+++ test/case/cli/system-output/run/rauc-installation-status
