+++ test/case/cli/system-output/run/rauc_status_--detailed_--output-format=json
